use long instead of short for {Ver,Desc}File size in pkgcache.h patch from Víctor Manuel Jáquez Leal, thanks! (Closes: #538917)

+++ apt-pkg/pkgcache.h
@@ -245,7 +245,7 @@ struct pkgCache::VerFile						/*{{{*/
    map_ptrloc File;           // PackageFile
    map_ptrloc NextFile;       // PkgVerFile
    map_ptrloc Offset;         // File offset
-   unsigned short Size;
+   unsigned long Size;
 };
 									/*}}}*/
 struct pkgCache::DescFile						/*{{{*/
@@ -253,7 +253,7 @@ struct pkgCache::DescFile						/*{{{*/
    map_ptrloc File;           // PackageFile
    map_ptrloc NextFile;       // PkgVerFile
    map_ptrloc Offset;         // File offset
-   unsigned short Size;
+   unsigned long Size;
 };
 									/*}}}*/
 struct pkgCache::Version						/*{{{*/
